+++ changelog.md
@@ -5,12 +5,20 @@
 - Create a helper for triangles.
 - Resize/bestfit seems to fail in Firefoxe's dark mode (plugin).
 - Extend demo 07-voronoi: "randomize" could be a smooth transition :)
+- Todo: Bounding box of PBText
+- Todo: Bounding box of PBImage
+- Todo: think about if Vector should implement `Intersectable` as well.
+- Todo: PBText shoould implement Bounds.
+- Todo: TBImage should implement Bounds.
+- Todo: check if BezierPathInteractionHelper can work without every possible redraw (event mouseIsOver)
 
 - 2025-05-07
   - Added `PBContentChangeListener` to PB to track if the drawable content changed.
   - Added `PlogBoilerplate.addContentChangeListener` and `PlogBoilerplate.removeContentChangeListener`.
   - Moving full vectors now by default when vector point a is moved.
   - (Demos) Adding a tiny content manager pane.
+  - Class `Vertex` is now implementing interface `IBounded` (to meet convention).
+  - Tweaking the demo `57-shape-reflecting-rays` to interact with the new content manager tool. Adding removing shapes is now possible.
 - 2025-05-05
   - Added optional params `draw` and `fill` to BezierPathInteractionHelper.drawHandleLines`method.
   - Demo 58-shape-reflecting-rays: Refactored`rebuildShapes`and moved to external funcion`createRandomShapes`.